Module Tezt_tx_kernel.Contracts

A set of helper functions for creating and managing L1 smart contracts that are used for interacting with the tx-kernel.

val prepare_mint_and_deposit_contract : ?hooks:Tezt.Process.hooks -> Tezt_tezos.Client.t -> Tezt_tezos.Protocol.t -> Tezos_protocol_alpha.Protocol.Contract_hash.t Lwt.t

Originate a contract that will mint and transfer tickets to the tx kernel. Uses the contract mini_scenarios/smart_rollup_mint_and_deposit_ticket.tz.

val prepare_receive_withdrawn_tickets_contract : ?hooks:Tezt.Process.hooks -> Tezt_tezos.Client.t -> Tezt_tezos.Protocol.t -> Tezos_protocol_alpha.Protocol.Contract_hash.t Lwt.t

Originate ticket receiver contract that will receive withdrawals. Uses the contract mini_scenarios/smart_rollup_receive_tickets.tz.

val deposit_string_tickets : ?hooks:Tezt.Process.hooks -> Tezt_tezos.Client.t -> mint_and_deposit_contract:string -> sc_rollup_address:string -> destination_l2_addr:string -> ticket_content:string -> amount:int -> unit Lwt.t

deposit_string_tickets ?hooks client ~sc_rollup_address ~ticket_content ~mint_and_deposit_contract ~destination_l2_addr ~amount deposits a string ticket into a rollup.

parameter client
The L1 client.

parameter mint_and_deposit_contract
The contract for minting and depositing tickets. This contract can be created by calling prepare_receive_withdrawn_tickets_contract.

parameter sc_rollup_address
The address of the smart-rollup to which we deposit the ticket.

parameter destination_l2_addr
The L2 address to which we deposit the ticket.

parameter ticket_content
The content of the ticket we deposit.

parameter amount
The amount of the ticket we deposit.