Module Tezt_tx_kernel.Transaction_batch

Module for constructing an encoded batch of transactions. The transactions can be sent to DAC/DAL as a payload or can be wrapped in an external message frame and sent to the L1 shared inbox.

The un-wrapped version corresponds to: inbox::external::external::v1::ParsedBatch

The version wrapped by the external message frame corresponds to: kernel_sdk::encoding::src::inbox::ExternalMessageFrame

type t

A batch of transactions.

val empty : t

An empty batch.

val add_transfer : counter:int -> signer_secret_key:Tezos_crypto.Signature.Ed25519.Secret_key.t -> ?signer:Types.signer -> destination:Tezos_crypto.Signature.Ed25519.Public_key_hash.t -> ticketer:string -> ticket_content:string -> amount:int -> t -> t

add_transfer ~counter ~signer_secret_key ?signer ~destination ~ticketer ~ticket_content ~amount t adds a ticket transfer operation to the batch.

parameter counter
The transaction counter that should increase by one with each transaction executed by the signer's account.

parameter signer_secret_key
The secret key of the sender. Used for signing the operation.

parameter signer
Allows to manually set the signer of the operation, e.g. to craft an operation with a wrong signature, or use the signer's public key. If omitted, the signer will be the public key hash derived from the signer_secret_key.

parameter destination
The L2 account in which we send the ticket to.

parameter ticketer
The ticketer of the ticket that is being transferred.

parameter ticket_content
The ticket content of the ticket that is being transferred.

parameter amount
The number of tickets to transfer.

val add_withdraw : counter:int -> signer_secret_key:Tezos_crypto.Signature.Ed25519.Secret_key.t -> ?signer:Types.signer -> destination:Tezos_protocol_alpha.Protocol.Contract_hash.t -> entrypoint:string -> ticketer:string -> ticket_content:string -> amount:int -> t -> t

add_withdraw ~counter ~signer_secret_key ?signer ~destination ~entrypoint ~ticketer ~ticket_content ~amount t adds a ticket withdraw operation to the batch.

parameter counter
The transaction counter that should increase by one with each transaction executed by the signer's account.

parameter signer_secret_key
The secret key of the withdrawer. Used for signing the operation.

parameter signer
Allows to manually set the signer of the operation, e.g. to craft an operation with a wrong signature, or use the signer's public key. If omitted, the signer will be the public key hash derived from the signer_secret_key.

parameter destination
The L1 contract in which we send the ticket to.

parameter entrypoint
The L1 contract's entrypoint in which we send the ticket to.

parameter ticketer
The ticketer of the ticket that is being withdrawn.

parameter ticket_content
The ticket content of the ticket that is being withdrawn.

parameter amount
The number of tickets to withdraw.

val make_encoded_batch : ?wrap_with: [ `External_message_frame of Tezos_crypto.Hashed.Smart_rollup_address.t | `None ] -> t -> string

make_encoded_batch ?wrap_with t constructs the encoded transaction batch. When wrap_with = `External_message_frame sc_rollup_addr, it wraps the encoded batch with an external message frame so it can be sent to the L1 inbox. When wrap_with = `None, it does not wrap the encoded batch. This is useful when we want to transmit the batch via methods outside of the inbox, like DAC or DAL.

Note that the transactions in the encoded batch will appear in the order they were added. E.g. If you add operations A, B, then C, the batch appears as [A; B; C].

Raises Invalid_argument if you pass wrap_with = `External_message_frame sc_rollup_addr and the resulting encoded batch exceeds the external message limit.